refactor : LottoService 삭제
