+++ src/main/java/lotto/domain/WinLotto.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import java.util.List;
+
 public class WinLotto {
     private final Numbers numbers;
     private final Integer bonusNumber;
@@ -11,6 +13,10 @@ public class WinLotto {
     public WinLotto(Numbers numbers, Integer bonusNumber) {
         this.numbers = numbers;
         this.bonusNumber = bonusNumber;
+    }
+
+    public WinLotto(List<Integer> numbers, Integer bonusNumber) {
+        this(new Numbers(numbers), bonusNumber);
     }
 
     public Numbers numbers() {
